requirement.txt
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
faiss-cpu>=1.7.4  # or faiss-gpu if you have CUDA
python-dotenv>=1.0.0
